Идемпотентность стартовых скриптов
Добавлена проверка, что пользователь по умолчанию в контейнере уже изменён на пользователя заданного при старте контейнера

+++ postgresql-11/src/main/docker/etc/my_init.d/08_pre_setup_user.sh
@@ -2,6 +2,6 @@
 
 # Удаляем пользователя, пришедшего вместе с пакетом установки PostgreSQL.
 # Вместо него будет переименован пользователь system
-userdel -r postgres 2> /dev/null
+[[ "$(id -u system > /dev/null 2>&1; echo $?)" == "0" ]] && userdel -r postgres 2> /dev/null || true
 # Форсируем имя пользователя, т.к. для PostgreSQL требуется именно это имя.
 export PUSER="postgres"
